lab12/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
docutils==0.20.1
et-xmlfile==1.1.0
idna==3.6
Kivy==2.3.0
Kivy-Garden==0.1.5
lxml==5.2.1
openpyxl==3.1.2
Pygments==2.17.2
python-docx==1.1.0
requests==2.31.0
setuptools==69.2.0
typing_extensions==4.10.0
urllib3==2.2.1
wheel==0.43.0
